+++ client/src/common/utils/timeConstants.js
@@ -1,0 +1,24 @@
+/**
+ * millis to seconds
+ * @type {number}
+ */
+export const mts = 1000;
+
+/**
+ * millis to minutes
+ * @type {number}
+ */
+export const mtm = 1000 * 60;
+
+
+/**
+ * millis to hours
+ * @type {number}
+ */
+export const mth = 1000 * 60 * 60;
+
+/**
+ * milliseconds in a day
+ * @type {number}
+ */
+export const DAY_TO_MS = 86400000;
