+++ server/src/classes/timer/__tests__/classUtils.test.js
