+++ server/src/classes/timer/__tests__/timer.test.js
