+++ server/src/classes/timer/classUtils.js
